requirements.txt
asgi-redis==0.14.1
asgiref==0.14.0
autobahn==0.16.0
channels==0.17.2
daphne==0.15.0
dj-database-url==0.4.0
dj-static==0.0.6
django==1.9.5
django-colorful==1.2
django-easy-select2==1.3.3
django-material>=0.9.0
django-redis-cache==1.7.0
django-smart-selects>=1.2.2
django-widget-tweaks==1.4.1
djangorestframework==3.4.7
futures==3.0.5
gunicorn==19.4.5
msgpack-python==0.4.8
Pillow==3.3.1
ply==3.9
psycopg2==2.6.2
redis==2.10.5
six==1.10.0
slimit==0.8.1
static3==0.7.0
Twisted==16.4.1
txaio==2.5.1
whitenoise==2.0.6
zope.interface==4.3.2
dj-database-url==0.4.0
dj-static==0.0.6
gunicorn==19.4.5
whitenoise==2.0.6
